Move fopen wrapper to separate file
